testPythonPkg reference counts docker active docker containers before trying to free them

+++ src/org/ds/python/toxFunctions.groovy
@@ -30,6 +30,7 @@ def getToxTestsParallel(args = [:]){
     def preRunClosure = args['beforeRunning']
     def retries = args.containsKey('retry') ? args.retry : 1
     def dockerRunArgs = args.get('dockerRunArgs', '')
+
     script{
         def envs
         def originalNodeLabel
@@ -57,16 +58,21 @@ def getToxTestsParallel(args = [:]){
             }
         }
         echo "Found tox environments for ${envs.join(', ')}"
+        def hosts = [:]
         def jobs = envs.collectEntries({ tox_env ->
             def jenkinsStageName = "${envNamePrefix} ${tox_env}"
             [jenkinsStageName,{
                 retry(retries){
                     node(label){
+                        if (!hosts.containsKey(env.NODE_NAME)){
+                            hosts[env.NODE_NAME] = 0
+                        }
                         ws{
                             checkout scm
                             def dockerImageForTesting = docker.build(dockerImageName, "-f ${dockerfile} ${dockerArgs} . ")
                             try{
                                 dockerImageForTesting.inside(dockerRunArgs){
+                                    hosts[env.NODE_NAME] += 1
                                     if(preRunClosure != null){
                                         preRunClosure()
                                     }
@@ -89,22 +95,29 @@ def getToxTestsParallel(args = [:]){
                                     )
                                 }
                             } finally {
-                                if(isUnix()){
-                                    def runningContainers = sh(
-                                                               script: "docker ps --no-trunc --filter ancestor=${dockerImageForTesting.imageName()} --format {{.Names}}",
-                                                               returnStdout: true,
-                                                               )
-                                    if (!runningContainers?.trim()) {
-                                        sh(
-                                            label: "Untagging Docker Image used to run tox",
-                                            script: "docker image rm --no-prune ${dockerImageForTesting.imageName()}",
-                                            returnStatus: true
-                                        )
+                                hosts[env.NODE_NAME] -= 1
+                                if (hosts[env.NODE_NAME] == 0){
+                                    if(isUnix()){
+                                        def runningContainers = sh(
+                                                                   script: "docker ps --no-trunc --filter ancestor=${dockerImageForTesting.imageName()} --format {{.Names}}",
+                                                                   returnStdout: true,
+                                                                   )
+                                        if (!runningContainers?.trim()) {
+                                            sh(
+                                                label: "Untagging Docker Image used to run tox",
+                                                script: "docker image rm --no-prune ${dockerImageForTesting.imageName()}",
+                                                returnStatus: true
+                                            )
+                                        }
+                                        sh(script: "docker ps --no-trunc --filter ancestor=${dockerImageForTesting.imageName()} --format {{.Names}}", returnStatus: true)
+                                    } else {
+                                        def powershellScript ="${env.WORKSPACE_TMP}/cleanupUnusedDockerImage.ps1"
+                                        if (! fileExists(powershellScript) ){
+                                            def scriptContent = libraryResource 'org/ds/cleanupUnusedDockerImage.ps1'
+                                            writeFile file: powershellScript, text: scriptContent
+                                        }
+                                        powershell(script: "${powershellScript} ${dockerImageForTesting.imageName()}", returnStatus: true)
                                     }
-                                    sh(script: "docker ps --no-trunc --filter ancestor=${dockerImageForTesting.imageName()} --format {{.Names}}", returnStatus: true)
-                                } else {
-                                    def powershellScript = libraryResource 'com/ds/cleanupUnusedDockerImage.ps1'
-                                    powershell(script: "${powershellScript} ${dockerImageForTesting.imageName()}", returnStatus: true)
                                 }
                             }
                         }
